METAMODEL-1166: Moved classes to more appropriate packages
