+++ core/src/main/java/org/apache/metamodel/membrane/controllers/TableDataController.java
@@ -33,9 +33,9 @@ import org.apache.metamodel.data.WhereClauseBuilder;
 import org.apache.metamodel.delete.RowDeletionBuilder;
 import org.apache.metamodel.insert.RowInsertionBuilder;
 import org.apache.metamodel.membrane.app.DataContextTraverser;
-import org.apache.metamodel.membrane.app.TenantContext;
-import org.apache.metamodel.membrane.app.TenantRegistry;
 import org.apache.metamodel.membrane.app.config.JacksonConfig;
+import org.apache.metamodel.membrane.app.registry.TenantContext;
+import org.apache.metamodel.membrane.app.registry.TenantRegistry;
 import org.apache.metamodel.membrane.swagger.model.Operator;
 import org.apache.metamodel.membrane.swagger.model.PostDataRequest;
 import org.apache.metamodel.membrane.swagger.model.PostDataRequestDelete;
